updated to TeX Live 2016.

+++ texk/libdpx/t1_char.h
@@ -15,10 +15,10 @@ typedef struct {
   } seac;
 } t1_ginfo;
 
-extern int  t1char_get_metrics (card8 *src, long srclen,
+extern int  t1char_get_metrics (card8 *src, int srclen,
 				cff_index *subrs, t1_ginfo *ginfo);
-extern long t1char_convert_charstring (card8 *dst, long dstlen,
-				       card8 *src, long srclen,
+extern int  t1char_convert_charstring (card8 *dst, int dstlen,
+				       card8 *src, int srclen,
 				       cff_index *subrs,
 				       double default_width, double nominal_width,
 				       t1_ginfo *ginfo);
